Implemented a way to modify the bandwidth while the simulation is running
Implemented a way to modify the forward and return links bandwidth during simulation.

The general idea is the following : 
The Gateway (GW), the Satellite Terminals (ST), and Satellite (SAT) listen for connections on the port 5335.
When a message is received, the System (GW, ST or SAT) calls a function of ConfUpdateXMLParser to update the "/etc/opensand/core_global.conf" configuration file (which contains the bandwidth used for the forward and return links).
Then, the whole configuration is unloaded (call to Conf::unloadConfig() defined in opensand-conf/src/Configuration.cpp), and reloaded.
This is done in order to use the new parameters defined in the configuration file.
Finally, the functions that used the modified value of the bandwidth are called again to reinitialize them correctly (this part might be dirty as my comprehension of OpenSand is not complete, but seems to do the job as far as i understand).

The "listening" on port 5335 is implemented in the Dvb blocks of the ST (BlockDvbTal.cpp), of the SAT (BlockDvbSat.cpp) and of the gateway (BlockDvbNcc.cpp).
The request handling and configuration update is done in BlockDvbTal.cpp (for the ST), in BlockDvbSAT.cpp (for the SAT), and in BlockDvbNcc.cpp, SpotDownward.cpp, SpotUpwardTransp.cpp, and  SpotUpwardRegen.cpp (for the GW).