Added GW ID check before executing a ConfUpdateRequest in ST
Added a request->GW_id check before executing a ConfUpdateRequest in ST, to avoid a ST executing a request that do not concern it.

Before this fix, if you had the following topology : 

GW0, GW6
ST1 in SPOT1, managed by GW0
ST2 in SPOT1, managed by GW6

And you sent a request saying "change Bandwidth of SPOT1, GW0 to 50Mhz"
Before this fix, ST1 and ST2 would execute the request (since it was for their spot (SPOT1), and no GW id check was made).

After this fix, only ST1 will execute it (as intended).